requirements.txt
tensorflow==2.11.0
numpy==1.23.5
opencv-python==4.7.0.72
pillow==9.3.0
matplotlib==3.6.2
streamlit==1.16.0
keras==2.11.0
